drivers:platform:stm32:dma: Fix DMA Abort API
The DMA_Abort_IT disables the DMA and rest of the flags are cleared
in the ISR. At this point, the execution do not reach ISR since the
IRQ is disabled in no_os_dma_abort platform's abort is called.
Hence abort needs to be done without ISR branch.

HAL returns an error when application tries to abort an already
completed DMA transaction.

Signed-off-by: Naga Himanshu Indraganti <naga.indraganti@analog.com>